db.py
import mysql.connector

def get_db():
    return mysql.connector.connect(
        host="localhost",
        user="root",
        password="", # Default XAMPP/WAMP password is empty
        database="online_exam"
    )
