requirements.txt
flask
mysql-connector-python
pandas
numpy
scikit-learn
joblib
